+++ src/App.css
@@ -1,6 +1,7 @@
 
 @import url('https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200');
 
+@import 'bootstrap/dist/css/bootstrap.min.css';
 .App {
   text-align: center;
 }
