+++ src/App.js
@@ -1,4 +1,5 @@
 import EstimationTool from "./pages/view/home";
+
 import './App.css';
 
 function App() {
